fix: speed up sam init help text by 100ms (#4213)
- make `sam init` follow convention by having a separate command.py
  file, and ensuring that `__init__.py` only imports cli.
