+++ tests/unit/commands/init/test_non_interactive_mode.py
@@ -3,12 +3,12 @@ from unittest.mock import Mock, patch
 
 import click
 
-from samcli.commands.init import non_interactive_validation
+from samcli.commands.init.command import non_interactive_validation
 from samcli.lib.utils.packagetype import IMAGE, ZIP
 
 
 class TestNonInteractiveMode(TestCase):
-    @patch("samcli.commands.init.click.get_current_context")
+    @patch("samcli.commands.init.command.click.get_current_context")
     def test_interactive_mode(self, mocked_ctx):
         mocked_ctx.return_value = Mock(
             params={
@@ -22,7 +22,7 @@ class TestNonInteractiveMode(TestCase):
 
         mocked_func.assert_called_once()
 
-    @patch("samcli.commands.init.click.get_current_context")
+    @patch("samcli.commands.init.command.click.get_current_context")
     def test_non_interactive_mode_with_non_java_image(self, mocked_ctx):
         mocked_ctx.return_value = Mock(
             params={"no_interactive": True, "package_type": IMAGE, "base_image": "amazon/dotnet5.0"}
@@ -34,7 +34,7 @@ class TestNonInteractiveMode(TestCase):
 
         mocked_func.assert_called_once()
 
-    @patch("samcli.commands.init.click.get_current_context")
+    @patch("samcli.commands.init.command.click.get_current_context")
     def test_non_interactive_mode_with_java_zip_dependency_manager(self, mocked_ctx):
         mocked_ctx.return_value = Mock(
             params={"no_interactive": True, "package_type": ZIP, "dependency_manager": "maven"}
@@ -45,7 +45,7 @@ class TestNonInteractiveMode(TestCase):
         wrapped_func()
         mocked_func.assert_called_once()
 
-    @patch("samcli.commands.init.click.get_current_context")
+    @patch("samcli.commands.init.command.click.get_current_context")
     def test_non_interactive_mode_with_java_image_dependency_manager(self, mocked_ctx):
         mocked_ctx.return_value = Mock(
             params={"no_interactive": True, "package_type": IMAGE, "dependency_manager": "maven"}
@@ -56,7 +56,7 @@ class TestNonInteractiveMode(TestCase):
         wrapped_func()
         mocked_func.assert_called_once()
 
-    @patch("samcli.commands.init.click.get_current_context")
+    @patch("samcli.commands.init.command.click.get_current_context")
     def test_non_interactive_mode_with_java_image_no_dependency_manager(self, mocked_ctx):
         mocked_ctx.return_value = Mock(
             params={"no_interactive": True, "package_type": IMAGE, "base_image": "amazon/java8"}
@@ -68,7 +68,7 @@ class TestNonInteractiveMode(TestCase):
             wrapped_func()
         mocked_func.assert_not_called()
 
-    @patch("samcli.commands.init.click.get_current_context")
+    @patch("samcli.commands.init.command.click.get_current_context")
     def test_non_interactive_mode_with_zip_no_dependency_manager(self, mocked_ctx):
         mocked_ctx.return_value = Mock(params={"no_interactive": True, "package_type": ZIP})
         mocked_func = Mock()
